Redoes the setup of the inventory page so that the main functionality is inside of a component instead of the page itself.
