+++ reactApplication/react_next_app/src/app/inventory_page/inventory_page_stylesheet.css
@@ -1,8 +1,3 @@
 .inventory-main-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100vh;
-    padding: 20px;
     margin-top: 80px;
   }
