+++ reactApplication/react_next_app/src/components/inventory/inventory.css
@@ -1,0 +1,7 @@
+.inventory_main_container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100vh;
+    padding: 20px;
+  }
